move clickOutside to own file
